+++ css/package.json
@@ -22,10 +22,10 @@
       "require": "./dist/colors.cjs.js",
       "types": "./dist/colors.d.ts"
     },
-    "./dist/windi-keep-rollup-plugin": {
-      "import": "./dist/windi-keep-rollup-plugin.es.mjs",
-      "require": "./dist/windi-keep-rollup-plugin.cjs.js",
-      "types": "./dist/windi-keep-rollup-plugin.d.ts"
+    "./dist/color-constants": {
+      "import": "./dist/color-constants.es.mjs",
+      "require": "./dist/color-constants.cjs.js",
+      "types": "./dist/color-constants.d.ts"
     }
   },
   "scripts": {
