+++ css/src/icon-extractor-tools.ts
@@ -1,16 +1,11 @@
 import _ from 'lodash'
-import { COLOR_PREFIXES, colors } from './colors'
+import { colors } from './colors'
+import {
+  COLOR_PREFIXES,
+  ICON_ATTRIBUTE_NAMES_TO_CLASS_GENERATOR_ROOT,
+} from './color-constants'
 
 const { camelCase, kebabCase } = _
-
-const ICON_ATTRIBUTE_NAMES_TO_CLASS_GENERATOR_ROOT = {
-  FillColor: (attrValue: string) => `icon-light-${attrValue}`,
-  StrokeColor: (attrValue: string) => `icon-dark-${attrValue}`,
-  SecondaryFillColor: (attrValue: string) =>
-    `icon-light-secondary-${attrValue}`,
-  SecondaryStrokeColor: (attrValue: string) =>
-    `icon-dark-secondary-${attrValue}`,
-} as const
 
 const ICON_ATTRIBUTE_NAMES_TO_CLASS_GENERATOR: Record<
   string,
